requirements.txt
Django>=5.1.3,<5.2
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
pytest-xdist==3.6.1
pytest-sugar==1.0.0
python-dotenv>=1.0.1,<1.1
django-crispy-forms>=2.3,<2.4
crispy-bootstrap5>=2024.10,<2024.11
requests>=2.32.3,<2.33